Sort scedules of new trains.

+++ backend/src/test/java/com/planifcarbon/backend/model/StationTerminusTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 public class StationTerminusTest {
     @ParameterizedTest
     @MethodSource("generateData")
-    public void testStation(String name, double la, double lo, List<Long> schedules) {
+    public void testStation(String name, double la, double lo, List<Integer> schedules) {
         StationTerminus s = new StationTerminus(name, la, lo, schedules);
         assertEquals(name, s.getName());
         assertEquals(la, s.getCoordinates().getLatitude());
@@ -23,7 +23,7 @@ public class StationTerminusTest {
 
     @ParameterizedTest
     @MethodSource("generateData")
-    public void testIsInMetro(String name, double la, double lo, List<Long> schedules) {
+    public void testIsInMetro(String name, double la, double lo, List<Integer> schedules) {
         StationTerminus s = new StationTerminus(name, la, lo, schedules);
         assertTrue(s.isInMetro());
     }
@@ -35,5 +35,5 @@ public class StationTerminusTest {
         });
     }
 
-    static Stream<Arguments> generateData() { return Stream.of(Arguments.of("ST1", 1, 2, List.of(1l, 2l))); }
+    static Stream<Arguments> generateData() { return Stream.of(Arguments.of("ST1", 1, 2, List.of(1, 2))); }
 }
